Added the User and Entry models
The Entry and User models have been added along with tests for them. The test file seems to create a peewee.db file for some reason that needs to be deleted each time the test file is ran. As far as I can tell the test_journal.py file is the same as the one from the Tacocat project. I am even using the same virtuall environment.