Added a main/secondary entry view.
The webpage now displays all or only the user's journal entries. The webpage has had some interactivity in various places as well. Things like the delete button only show up on the page if that user is logged in. The test file seems messed up a little. If I use the current peewee.db file I seem to get an error. This is fixed though if I delete peeewee.db